Move rainmachine base entity to separate module
