+++ homeassistant/components/rainmachine/sensor.py
@@ -20,9 +20,9 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.util.dt import utc_from_timestamp, utcnow
 
-from . import RainMachineConfigEntry, RainMachineData, RainMachineEntity
+from . import RainMachineConfigEntry, RainMachineData
 from .const import DATA_PROGRAMS, DATA_PROVISION_SETTINGS, DATA_ZONES
-from .model import RainMachineEntityDescription
+from .entity import RainMachineEntity, RainMachineEntityDescription
 from .util import (
     RUN_STATE_MAP,
     EntityDomainReplacementStrategy,
